+++ R/model_eval.R
@@ -51,11 +51,11 @@ model_eval <- function(mod, data=NULL, ..., skeleton=FALSE, ncont=3,
        training_data <- data_from_tilde(get_training_data(mod),
                                         formula_from_mod(mod))
        eval_data <- training_data
+       names(eval_data)[1] <- ".response"
        response_in_data <- TRUE
      }
   } else {
     eval_data <- training_data <- data
-    names(training_data)[1] <- ".response"
     # the argument data might or might not have the response name
     response_in_data <- response_var_name %in% names(data)
   }
@@ -81,7 +81,7 @@ model_eval <- function(mod, data=NULL, ..., skeleton=FALSE, ncont=3,
 
   if (response_in_data) {
     Residuals <- data.frame(.resid = eval_data[[response_var_name]] - Fitted$.output)
-    names(training_data)[1] <- ".response" # give it a generic name
+    names(training_data)[names(training_data) == response_var_name] <- ".response" # give it a generic name
     return(dplyr::bind_cols(training_data, Fitted, Residuals))
   } else {
     return(dplyr::bind_cols(eval_data, Fitted))